chore: add new types and interfaces for address management

+++ web3scanner.go
@@ -15,11 +15,19 @@ import (
 	"github.com/qiaopengjun5162/web3scanner/database"
 )
 
+// Web3Scanner 是一个结构体，用于扫描和监控Web3相关的活动或数据。
+// 它包含数据库连接和 shutdown、stopped 两个字段，用于控制扫描器的停止和检查停止状态。
 type Web3Scanner struct {
+	// db 是一个数据库连接实例，用于执行数据库操作。
 	db *database.DB
 
+	// shutdown 是一个context.CancelCauseFunc类型的函数，
+	// 用于在需要停止扫描器时调用，以优雅地关闭扫描器。
 	shutdown context.CancelCauseFunc
-	stopped  atomic.Bool
+
+	// stopped 是一个原子布尔值，用于表示扫描器是否已经停止。
+	// 这提供了一种线程安全的方式来检查扫描器的停止状态。
+	stopped atomic.Bool
 }
 
 // NewWeb3Scanner creates a new instance of Web3Scanner.
@@ -58,7 +66,7 @@ func (ws *Web3Scanner) Start(_ context.Context) error {
 		Address:     common.HexToAddress("0x0fa09C3A328792253f8dee7116848723b72a6d2e"),
 		AddressType: 1,
 		PublicKey:   "0x0fa09C3A328792253f8dee7116848723b72a6d2e",
-		Timestamp:   uint64(time.Now().Unix()),
+		Timestamp:   time.Now().Unix(),
 	}
 	batchAddress = append(batchAddress, addressItem)
 	err := ws.db.Addresses.StoreAddresses(batchAddress)
